+++ examples/addressbookserver.php
@@ -37,7 +37,7 @@ $carddavBackend   = new Sabre\CardDAV\Backend\PDO($pdo);
 // Setting up the directory tree //
 $nodes = [
     new Sabre\DAVACL\PrincipalCollection($principalBackend),
-//    new Sabre\CalDAV\CalendarRootNode($authBackend, $caldavBackend),
+//    new Sabre\CalDAV\CalendarRoot($authBackend, $caldavBackend),
     new Sabre\CardDAV\AddressBookRoot($principalBackend, $carddavBackend),
 ];