+++ lib/DAVACL/Principal.php
@@ -246,13 +246,13 @@ class Principal extends DAV\Node implements IPrincipal, DAV\IProperties, IACL {
      */
     public function getACL() {
 
-        return array(
-            array(
+        return [
+            [
                 'privilege' => '{DAV:}read',
-                'principal' => $this->getPrincipalUrl(),
+                'principal' => '{DAV:}authenticated',
                 'protected' => true,
-            ),
-        );
+            ],
+        ];
 
     }