+++ tests/Sabre/CalDAV/Backend/AbstractPDOTest.php
@@ -333,7 +333,7 @@ abstract class AbstractPDOTest extends \PHPUnit_Framework_TestCase {
         $backend = new PDO($this->pdo);
         $returnedId = $backend->createCalendar('principals/user2','somerandomid',array());
 
-        $object = "BEGIN:VCALENDAR\r\nBEGIN:VEVENT\r\nDTSTART;VALUE=DATE-TIME:20120101T100000Z\r\nRRULE:FREQ=DAILY\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n";
+        $object = "BEGIN:VCALENDAR\r\nBEGIN:VEVENT\r\nDTSTART;VALUE=DATE-TIME:20120101T100000Z\r\nRRULE:FREQ=DAILY\r\nUID:foo\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n";
 
         $backend->createCalendarObject($returnedId, 'random-id', $object);
 
@@ -357,7 +357,7 @@ abstract class AbstractPDOTest extends \PHPUnit_Framework_TestCase {
         $backend = new PDO($this->pdo);
         $returnedId = $backend->createCalendar('principals/user2','somerandomid',array());
 
-        $object = "BEGIN:VCALENDAR\r\nBEGIN:VEVENT\r\nDTSTART;VALUE=DATE-TIME:20120101T100000Z\r\nDTEND;VALUE=DATE-TIME:20120101T110000Z\r\nRRULE:FREQ=DAILY;COUNT=1000\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n";
+        $object = "BEGIN:VCALENDAR\r\nBEGIN:VEVENT\r\nDTSTART;VALUE=DATE-TIME:20120101T100000Z\r\nDTEND;VALUE=DATE-TIME:20120101T110000Z\r\nUID:foo\r\nRRULE:FREQ=DAILY;COUNT=1000\r\nEND:VEVENT\r\nEND:VCALENDAR\r\n";
 
         $backend->createCalendarObject($returnedId, 'random-id', $object);
 
@@ -811,4 +811,45 @@ abstract class AbstractPDOTest extends \PHPUnit_Framework_TestCase {
         $subs = $backend->getSubscriptionsForUser('principals/user1');
         $this->assertEquals(0, count($subs));
     }
+
+    function testSchedulingMethods() {
+
+        $backend = new PDO($this->pdo);
+
+        $calData = "BEGIN:VCALENDAR\r\nEND:VCALENDAR\r\n";
+
+        $backend->createSchedulingObject(
+            'principals/user1',
+            'schedule1.ics',
+            $calData
+        );
+
+        $expected = [
+            'calendardata' => $calData,
+            'uri' => 'schedule1.ics',
+            'etag' => '"' . md5($calData) . '"',
+            'size' => strlen($calData)
+        ];
+
+        $result = $backend->getSchedulingObject('principals/user1', 'schedule1.ics');
+        foreach($expected as $k=>$v) {
+            $this->assertArrayHasKey($k, $result);
+            $this->assertEquals($v, $result[$k]);
+        }
+
+        $results = $backend->getSchedulingObjects('principals/user1');
+
+        $this->assertEquals(1, count($results));
+        $result = $results[0];
+        foreach($expected as $k=>$v) {
+            $this->assertEquals($v, $result[$k]);
+        }
+
+        $backend->deleteSchedulingObject('principals/user1', 'schedule1.ics');
+        $result = $backend->getSchedulingObject('principals/user1', 'schedule1.ics');
+
+        $this->assertNull($result);
+
+    }
+
 }