+++ tests/Sabre/CalDAV/CalendarHomeSharedCalendarsTest.php
@@ -8,7 +8,7 @@ require_once 'Sabre/CalDAV/TestUtil.php';
 
 /**
  */
-class UserCalendarsSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
+class CalendarHomeSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
 
     protected $backend;
 
@@ -34,7 +34,7 @@ class UserCalendarsSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
             array()
         );
 
-        return new UserCalendars($this->backend, array(
+        return new CalendarHome($this->backend, array(
             'uri' => 'principals/user1'
         ));
 
@@ -51,7 +51,7 @@ class UserCalendarsSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
 
         $instance = $this->getInstance();
         $children = $instance->getChildren();
-        $this->assertEquals(4, count($children));
+        $this->assertEquals(3, count($children));
 
         // Testing if we got all the objects back.
         $hasShareable = false;
@@ -67,9 +67,6 @@ class UserCalendarsSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
             if ($child instanceof ISharedCalendar) {
                 $hasShared = true;
             }
-            if ($child instanceof Schedule\IOutbox) {
-                $hasOutbox = true;
-            }
             if ($child instanceof Notifications\ICollection) {
                 $hasNotifications = true;
             }
@@ -77,7 +74,6 @@ class UserCalendarsSharedCalendarsTest extends \PHPUnit_Framework_TestCase {
         }
         if (!$hasShareable) $this->fail('Missing node!');
         if (!$hasShared) $this->fail('Missing node!');
-        if (!$hasOutbox) $this->fail('Missing node!');
         if (!$hasNotifications) $this->fail('Missing node!'); 
 
     }