+++ tests/Sabre/CalDAV/Principal/UserTest.php
@@ -103,7 +103,7 @@ class UserTest extends \PHPUnit_Framework_TestCase {
         $expected = array(
             array(
                 'privilege' => '{DAV:}read',
-                'principal' => 'principals/user',
+                'principal' => '{DAV:}authenticated',
                 'protected' => true,
             ),
             array(